preds.py loads all models' .csv.gz prediction files in module scope and so is quite slow to import
better have something as frequently used as the Models enum in a cheap-to-import module
