+++ models/bowsr/join_bowsr_preds.py
@@ -7,9 +7,8 @@ import pymatviz
 from pymatviz.enums import Key
 from tqdm import tqdm
 
-from matbench_discovery.data import DataFiles
+from matbench_discovery.data import DataFiles, Model
 from matbench_discovery.enums import Task
-from matbench_discovery.preds import Model
 
 __author__ = "Janosh Riebesell"
 __date__ = "2022-09-22"
