+++ models/bowsr/test_bowsr.py
@@ -14,9 +14,8 @@ from pymatviz.enums import Key
 from tqdm import tqdm
 
 from matbench_discovery import timestamp, today
-from matbench_discovery.data import DataFiles, as_dict_handler
+from matbench_discovery.data import DataFiles, Model, as_dict_handler
 from matbench_discovery.enums import Task
-from matbench_discovery.preds import Model
 from matbench_discovery.slurm import slurm_submit
 
 __author__ = "Janosh Riebesell"
