+++ models/chgnet/ctk_structure_viewer.py
@@ -3,7 +3,7 @@ from crystal_toolkit.helpers.utils import hook_up_fig_with_struct_viewer
 from pymatviz.enums import Key
 from pymatviz.utils import PLOTLY
 
-from matbench_discovery.preds import Model
+from matbench_discovery.data import Model
 
 __author__ = "Janosh Riebesell"
 __date__ = "2023-03-07"
