+++ models/mace/analyze_mace.py
@@ -9,9 +9,8 @@ from pymatviz.enums import Key
 
 from matbench_discovery import SITE_FIGS
 from matbench_discovery import plots as plots
-from matbench_discovery.data import df_wbm
+from matbench_discovery.data import Model, df_wbm
 from matbench_discovery.enums import MbdKey
-from matbench_discovery.preds import Model
 
 __author__ = "Janosh Riebesell"
 __date__ = "2023-07-23"
